osint-workbench-v1/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-workbench-v1"
version = "1.0.0"
description = "OSINT investigative workbench v1 baseline"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.3",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.1.3",
  "celery>=5.4.0",
  "redis>=5.0.8",
  "httpx>=0.27.0",
  "minio>=7.2.7",
  "neo4j>=5.23.1",
  "python-multipart>=0.0.9",
  "email-validator>=2.2.0"
]

[tool.setuptools]
packages = ["api", "worker", "shared"]
